contador

+++ resources/views/welcome.blade.php
@@ -106,7 +106,7 @@
                                                 </div>
                                                 <div class="col-md-4">
                                                     <div class="row style_card_divisao style_card_divisao_horizontal" style="height: 90%;">
-                                                        <div class="col-md-12 style_card_medio_conteudo">{{count($etapa->candidatos)/2}}</div>
+                                                        <div class="col-md-12 style_card_medio_conteudo">{{(int) count($etapa->candidatos)/2}}</div>
                                                         <div class="col-md-12 style_card_medio_legenda">PESSOAS CADASTRADAS NESTA FAIXA ETÁRIA</div>
                                                     </div>
                                                 </div>
@@ -138,7 +138,7 @@
                                                 </div>
                                                 <div class="col-md-4">
                                                     <div class="row style_card_divisao style_card_divisao_horizontal" style="height: 90%;">
-                                                        <div class="col-md-12 style_card_medio_conteudo">{{count($etapa->candidatos)/2}}</div>
+                                                        <div class="col-md-12 style_card_medio_conteudo">{{(int)count($etapa->candidatos)/2}}</div>
                                                         <div class="col-md-12 style_card_medio_legenda">PESSOAS CADASTRADAS NESTE PÚBLICO</div>
                                                     </div>
                                                 </div>
@@ -170,7 +170,7 @@
                                                 </div>
                                                 <div class="col-md-4">
                                                     <div class="row style_card_divisao style_card_divisao_horizontal" style="height: 90%;">
-                                                        <div class="col-md-12 style_card_medio_conteudo">{{count($etapa->candidatos)/2}}</div>
+                                                        <div class="col-md-12 style_card_medio_conteudo">{{(int)count($etapa->candidatos)/2}}</div>
                                                         <div class="col-md-12 style_card_medio_legenda">PESSOAS CADASTRADAS NESTE PÚBLICO</div>
                                                     </div>
                                                 </div>
@@ -230,7 +230,7 @@
                                 <div class="row">
                                     <div class="col-md-12">
                                         <div class="row">
-                                            <div class="col-md-12 style_card_menor_conteudo">{{$quantPessoasCadastradas/2}}</div>
+                                            <div class="col-md-12 style_card_menor_conteudo">{{(int)$quantPessoasCadastradas/2}}</div>
                                             <div class="col-md-12 style_card_menor_legenda">TOTAL</div>
                                         </div>
                                     </div>
